requirements.txt
SQLAlchemy== 2.0.34
requests== 2.31.0
python-dotenv== 1.0.1
psycopg2-binary== 2.9.9
Flask== 3.0.3
